fix: exports pre-defined animation effects
