+++ jest.config.js
@@ -1,4 +1,4 @@
 module.exports = {
   preset: "ts-jest",
-  collectCoverageFrom: ["src/*.(ts|tsx)", "!src/*.d.ts"],
+  collectCoverageFrom: ["src/*.(ts|tsx)", "!src/index.ts", "!src/*.d.ts"],
 };
